Refactor: Move from follow folder to user folder
