+++ src/main/java/com/gokimpark/instaclone/domain/follow/FollowRepository.java
@@ -1,6 +1,6 @@
 package com.gokimpark.instaclone.domain.follow;
 
-import com.gokimpark.instaclone.domain.follow.dto.UserSimpleInfoDto;
+import com.gokimpark.instaclone.domain.user.dto.UserSimpleInfoDto;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.data.jpa.repository.Query;
 import org.springframework.data.repository.query.Param;
@@ -15,10 +15,10 @@ public interface FollowRepository extends JpaRepository<Follow, Follow.PK> {
     Long countByToUser(Long userId);
     Long countByFromUser(Long userId);
 
-    @Query(value = "select new com.gokimpark.instaclone.domain.follow.dto.UserSimpleInfoDto(u.username, u.name) from Follow f INNER JOIN User u ON f.toUser = u.id where f.fromUser = :userId")
+    @Query(value = "select new com.gokimpark.instaclone.domain.user.dto.UserSimpleInfoDto(u.username, u.name) from Follow f INNER JOIN User u ON f.toUser = u.id where f.fromUser = :userId")
     List<UserSimpleInfoDto> findAllByFromUser(@Param("userId") Long userId);
 
-    @Query(value = "select new com.gokimpark.instaclone.domain.follow.dto.UserSimpleInfoDto(u.username, u.name) from Follow f JOIN User u ON f.fromUser = u.id where f.toUser = :userId")
+    @Query(value = "select new com.gokimpark.instaclone.domain.user.dto.UserSimpleInfoDto(u.username, u.name) from Follow f JOIN User u ON f.fromUser = u.id where f.toUser = :userId")
     List<UserSimpleInfoDto> findAllByToUser(@Param("userId") Long userId);
 
     void deleteAllByFromUser(Long userId);
